+++ src/sched/connection.h
@@ -1,4 +1,4 @@
-/* $Id: connection.h,v 1.123 2005/04/27 21:41:53 jonas Exp $ */
+/* $Id: connection.h,v 1.124 2005/04/27 22:12:06 jonas Exp $ */
 
 #ifndef EL__SCHED_CONNECTION_H
 #define EL__SCHED_CONNECTION_H
@@ -97,7 +97,7 @@ void set_connection_timeout(struct connection *);
  * that should be probably something else than data, but... ;-) */
 /* Returns 0 on success and -1 on failure. */
 int load_uri(struct uri *uri, struct uri *referrer, struct download *download,
-	     enum connection_priority pri, enum cache_mode cache_mode, int start);
+	     enum connection_priority pri, enum cache_mode cache_mode, off_t start);
 
 int is_entry_used(struct cache_entry *cached);
